Bringing in dev some stuff from the on-going patch 1.0.3

+++ sources/subs/Topic.subs.php
@@ -328,23 +328,27 @@ function removeTopics($topics, $decreasePostCount = true, $ignoreRecycling = fal
 	if (empty($messages))
 		$messages = messagesInTopics($topics);
 
-	// Remove all likes now that the topic is gone
-	$db->query('', '
-		DELETE FROM {db_prefix}message_likes
-		WHERE id_msg IN ({array_int:messages})',
-		array(
-			'messages' => $messages,
-		)
-	);
+	// If there are messages left in this topic
+	if (!empty($messages))
+	{
+		// Remove all likes now that the topic is gone
+		$db->query('', '
+			DELETE FROM {db_prefix}message_likes
+			WHERE id_msg IN ({array_int:messages})',
+			array(
+				'messages' => $messages,
+			)
+		);
 
-	// Remove all mentions now that the topic is gone
-	$db->query('', '
-		DELETE FROM {db_prefix}log_mentions
-		WHERE id_msg IN ({array_int:messages})',
-		array(
-			'messages' => $messages,
-		)
-	);
+		// Remove all mentions now that the topic is gone
+		$db->query('', '
+			DELETE FROM {db_prefix}log_mentions
+			WHERE id_msg IN ({array_int:messages})',
+			array(
+				'messages' => $messages,
+			)
+		);
+	}
 
 	// Delete messages in each topic.
 	$db->query('', '
